+++ src/SlepianCoeffsToSH.f95
@@ -51,10 +51,10 @@ subroutine SlepianCoeffsToSH(film, falpha, galpha, lmax, nmax, exitstatus)
 
     real(dp), intent(out) :: film(:,:,:)
     real(dp), intent(in) :: falpha(:), galpha(:,:)
-    integer, intent(in) :: lmax, nmax
-    integer, intent(out), optional :: exitstatus
+    integer(int32), intent(in) :: lmax, nmax
+    integer(int32), intent(out), optional :: exitstatus
     real(dp), allocatable :: f(:)
-    integer :: i, astat
+    integer(int32) :: i, astat
 
     if (present(exitstatus)) exitstatus = 0
